Make Deferred a wrapper for IO and provide after
Deferred now wraps IO in order to be able to pass data from the notification (now of type Notification (IO ()) to a deferred value to be observed after the SendT action has been executed. This allows to wait for notifications with a return value and makes it possible to implement a wider range of communication primitives in the SendT monad.