Add time tag only to leaf bundles
Given that the intention of the time tag is to start synths without jitter, and everything but the leaf bundles in the synchronisation tree is preparatory, only the leaf bundles need to be scheduled at the time tag, everything else needs to be execute immediately.

This refactoring has the nice side effect of removing the time argument to `whenDone', making the interface more modular.